ezROUND
A really ez [;)] function to round the results. Just save ezANOVA as an object and enter it to ezROUND(), there's only one argument, the number of digits.

Please consider this as an offer. If you're interested, I'd be willing to develop this function in a more flexible version that can handle all kinds of fits provided by the ez-Package.